Upgrade TabSidebar scope (now named as SideTabNavigation) & add missing props to ValueIndicator
