--- lib/components/peopleExperience/TabSidebar.js
@@ -1,20 +1,0 @@
-import { jsx as _jsx, jsxs as _jsxs } from "react/jsx-runtime";
-import { Stack, colors, Typography, ListItemText, List, styled, ListItemButton, } from '@mui/material';
-import { useTranslation } from 'react-i18next';
-const DECO_BORDER_WIDTH = '4px';
-const ListItem = styled(ListItemButton)(({ theme }) => ({
-    borderLeft: `${DECO_BORDER_WIDTH} solid transparent`,
-    transition: 'all 250ms',
-    '&.Mui-selected': {
-        color: theme.palette.primary.main,
-        borderLeftColor: theme.palette.primary.main,
-    },
-}));
-const TabSidebar = ({ selectedIndex, onChangeIndex, tabs, }) => {
-    const { t } = useTranslation();
-    const handleListItemClick = (_event, index) => {
-        onChangeIndex(index);
-    };
-    return (_jsxs(Stack, { sx: { gap: 2, py: 4, bgcolor: colors.grey[50], height: '100%' }, children: [_jsx(Typography, { variant: "h5", component: "h2", sx: { px: theme => `calc(${theme.spacing(2)} + ${DECO_BORDER_WIDTH})` }, children: t('RESULTS') }), _jsx(List, { component: "nav", "aria-label": "", sx: { display: 'flex', flexDirection: 'column', gap: 1 }, disablePadding: true, children: tabs.map(({ id, title }, index) => (_jsx(ListItem, { selected: selectedIndex === index, onClick: event => handleListItemClick(event, index), children: _jsx(ListItemText, { primary: title, primaryTypographyProps: { sx: { fontWeight: '500' } } }) }, id))) })] }));
-};
-export default TabSidebar;
