+++ lib/index.js
@@ -5,6 +5,7 @@ import LinearProgressWithLabel from './components/peopleExperience/LinearProgres
 import ListWithDivider from './components/peopleExperience/ListWithDivider';
 import { ParticipationItemSkeleton, ParticipationExpandableItem, ParticipationItem, } from './components/peopleExperience/Participation';
 import ResultContainer from './components/peopleExperience/ResultContainer';
+import SideTabNavigation from './components/peopleExperience/SideTabNavigation';
 import ValueIndicator from './components/peopleExperience/ValueIndicator';
 export * from '@mui/material';
 export * from '@mui/x-date-pickers';
@@ -42,4 +43,5 @@ export const PeopleExperience = {
     LinearProgressWithLabel,
     ActionsMenu,
     ActionsMenuItem,
+    TabSidebar: SideTabNavigation,
 };
