+++ src/components/peopleExperience/ValueIndicator.tsx
@@ -1,11 +1,20 @@
 import { InfoOutlined } from '@mui/icons-material';
-import { Skeleton, Stack, Tooltip, Typography } from '@mui/material';
+import {
+  Skeleton,
+  Stack,
+  StackProps,
+  Tooltip,
+  Typography,
+} from '@mui/material';
 
 type ValueIndicatorProps = {
   value: number | string;
   label: string;
   loading?: boolean;
   description?: string;
+  slotProps?: Partial<{
+    container: StackProps;
+  }>;
 };
 
 const ValueIndicator = ({
@@ -13,43 +22,50 @@ const ValueIndicator = ({
   label,
   loading = false,
   description,
-}: ValueIndicatorProps) => (
-  <Stack sx={{ alignItems: 'center', gap: 1 }}>
-    {loading && (
-      <Skeleton
-        variant="text"
-        width="2rem"
-      />
-    )}
-    {!loading && (
-      <Typography
-        variant="h5"
-        component="p"
-      >
-        {value.toLocaleString()}
-      </Typography>
-    )}
-    <Stack sx={{ flexDirection: 'row', gap: 1, alignItems: 'center' }}>
-      <Typography
-        variant="overline"
-        sx={{ color: theme => theme.palette.secondary.main }}
-      >
-        {label}
-      </Typography>
-      {description && (
-        <Tooltip
-          title={description}
-          placement="right-end"
-          slotProps={{ tooltip: { sx: { maxWidth: '135px' } } }}
-        >
-          <InfoOutlined
-            fontSize="small"
-            color="secondary"
-          />
-        </Tooltip>
+  slotProps = {},
+}: ValueIndicatorProps) => {
+  const { container } = slotProps;
+  return (
+    <Stack
+      {...container}
+      sx={{ alignItems: 'center', gap: 1, ...container?.sx }}
+    >
+      {loading && (
+        <Skeleton
+          variant="text"
+          width="2rem"
+        />
       )}
+      {!loading && (
+        <Typography
+          variant="h5"
+          component="p"
+        >
+          {value.toLocaleString()}
+        </Typography>
+      )}
+      <Stack sx={{ flexDirection: 'row', gap: 1, alignItems: 'center' }}>
+        <Typography
+          variant="overline"
+          sx={{ color: theme => theme.palette.secondary.main }}
+        >
+          {label}
+        </Typography>
+        {description && (
+          <Tooltip
+            title={description}
+            placement="right-end"
+            slotProps={{ tooltip: { sx: { maxWidth: '135px' } } }}
+          >
+            <InfoOutlined
+              fontSize="small"
+              color="secondary"
+            />
+          </Tooltip>
+        )}
+      </Stack>
     </Stack>
-  </Stack>
-);
+  );
+};
 
 export default ValueIndicator;
